Add favicons - oh the minefield
So it'd had been a while since I added favicons to a website, and boy
are there a lot of options these days. I followed this article somewhat
https://css-tricks.com/favicon-quiz/ and used realfavicongenerator.net
and also played around with favicomatic.com.

I decided not to use any that required manifests, presumably this won't
be seen much outside of Google Chrome in actuality.