Fix JS for IE11
This resolves a number of issues where this was failing in IE11. First
off the JS wasn't actually being transpiled as I wasn't using the
babel-loader for webpack - oops! I hadn't realised ES6 browser support
was so good in modern browsers. Secondly for Babel 7.3 you need a Symbol
and Symbol iterable polyfill so I installed core-js and then replaced
the promise polyfill with the one from that. Finally I needed a FormData
polyfill so I could use FormData.entries.