Fix missing padding and default visibility of rsvp-thanks
Using the hidden HTML attribute will mean this is hidden by default and
can still be overridden by CSS to show it - meaning the naked version of
this page doesn't show this extra text.

The margin also went wrong on the paragraph due to identical specificity
of class names. I resolved this by putting this on a parent element.