Rename images for better optimisation techniques
This renames the images so that it's clear whether they are for 2x or 3x
pixel densities. It then applies the image optimisation so that 2x
images have a quality of 50% and 3x have a quality of 33%.

This applies a technique from https://jakearchibald.com/2021/serving-sharp-images-to-high-density-screens/
of being much more leniant with image qualities at higher pixel
densities.