Parallel columns for "How to get there"
This uses columns at large size so each of the options are presented as
a column rather than a vertical linear flow.

I hit a nasty tachyons bug applying this, where the width 960px was in
both medium media queries and large media queries [1]. I've had to do
some rather nasty CSS to get past that.

[1]: https://github.com/tachyons-css/tachyons/issues/578