Hack: Fix off centre splash in IOS Safari
IOS Safari doesn't include the browser controls in 100vh which means
that something vertically centred looks off-centred on initial page
load, which is a bit of a bummer for our splash logo.

This applies a technique from https://css-tricks.com/css-fix-for-100vh-in-mobile-webkit/
to use the -webkit-fill-available to set the height.

I found that Safari removed the padding of the element from the height
of the -webkit-fill-available element so I had to move the padding to a
nested element.

I tried to do this with a fill-available CSS attribute without a
`@supports` hack however I found this broke the height in Chrome
browsers. So it seemed less risky to just apply this to IOS Safari.

I expect this same problem will affect other mobile browsers (since it
seems all vendors are aligning around Safari's 100vh understanding) but
it seemed to look ok in all the other mobile browsers I tested.

Hopefully IOS Safari 15 doesn't break all this.