Cope with differing content-type header casing
Annoyingly there's not a consistent way (I know of) to access these
proxied headers in the Lambda. I found that in my local environment I
received title cased headers (unmodified) but on AWS they were
lowercase.

This code is set-up to handle both, with the hope that there is no other
exotic combinations that can occur.