Add JS for Gifts form
This adds the ability to interact with the form in a neater way to using
anchors; validation of the form and the ability to repopulate the form
fields by query string. If any users click "Back" on the Stripe form
it'll take them to the form with a query string, so this will give them
back their details.