Reference AWS Secret Manager for env var
This uses AWS Secret Manager as a means to pull in a secret. This one
line of code may not look like much, but oh my gosh what a pain it was
to get here.

First idea: don't use any AWS secret/parameter services and just set the
secret as a cloud formation parameter. This approach worked but left the
secret in plain text in the samconfig.toml file, rejected.

Second idea: use parameter store. I created a ssm-secure parameter and
went to try reference it. Turns out you can't reference an SSM secure
parameter in a lambda cloud resource [1].

Third idea: use parameter store but a non-secure parameter. This worked
but I wasn't sure it was cool or not to use a non-secure parameter

Fourth idea: pony up the 40c a month and use AWS secret manager.

[1]: https://stackoverflow.com/questions/62305957/ssm-secure-reference-is-not-supported-in-awsec2instance-metadata